feat: use both set of annotations for request types
