--- servers/graphql-kotlin-server/src/benchmarks/kotlin/testtypes/GraphQLServerError.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright 2021 Expedia, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.expediagroup.graphql.server.testtypes
-
-import com.expediagroup.graphql.server.types.serializers.AnyNullableKSerializer
-import com.expediagroup.graphql.server.types.serializers.GraphQLErrorPathKSerializer
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-import com.fasterxml.jackson.annotation.JsonInclude
-import kotlinx.serialization.Serializable
-
-/**
- * GraphQL error representation that is spec complaint with serialization and deserialization.
- *
- * @see [GraphQL Specification](http://spec.graphql.org/June2018/#sec-Errors) for additional details
- */
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@Serializable
-data class GraphQLServerError(
-    val message: String,
-    val locations: List<GraphQLSourceLocation>? = null,
-    val path: List<@Serializable(with = GraphQLErrorPathKSerializer::class) Any>? = null,
-    val extensions: Map<String, @Serializable(with = AnyNullableKSerializer::class) Any?>? = null
-)
