+++ servers/graphql-kotlin-server/src/benchmarks/kotlin/testtypes/GraphQLServerRequest.kt
@@ -40,16 +40,10 @@ import kotlinx.serialization.json.JsonDecoder
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.decodeFromJsonElement
 
-/**
- * GraphQL server request abstraction that provides a convenient way to handle both single and batch requests.
- */
 @JsonDeserialize(using = GraphQLServerRequestDeserializer::class)
 @Serializable(with = GraphQLServerRequestKSerializer::class)
 sealed class GraphQLServerRequest
 
-/**
- * Wrapper that holds single GraphQLRequest to be processed within an HTTP request.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonDeserialize(using = JsonDeserializer.None::class)
@@ -61,9 +55,6 @@ data class GraphQLRequest(
     val extensions: Map<String, @Serializable(with = AnyNullableKSerializer::class) Any?>? = null
 ) : GraphQLServerRequest()
 
-/**
- * Wrapper that holds list of GraphQLRequests to be processed together within a single HTTP request.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonDeserialize(using = JsonDeserializer.None::class)
